Made assets service optional and removed provider from module (#96)
* Made assets service optional and removed provider from module

Providing `SkyAppAssetsService` caused issues for lazy-loaded routes because it would supersede the implemented assets service in the root app module. I made the provider optional and removed the erroneous provider from `SkyI18nModule` to fix it. I didn't have a way to test the issue that prompted the original change to provide the assets service, but the description of the issue in that PR doesn't seem like a valid issue to begin with. I guess we'll see.

* Removed unnecessary `beforeEach()`

* Fixed test description