fix(components/data-manager): assign specific types to properties of `SkyDataViewConfig`
Assign more specific types for the properties `additionalOptions`, `onClearAllClick`, and
`onSelectAllClick` of `SkyDataViewConfig`. This was done to address the
`@typescript-eslint/ban-types` ESLint rule. Marking as a breaking change, but this will likely not
cause any downstream effects.

BREAKING CHANGE: The properties `additionalOptions`, `onClearAllClick`, and `onSelectAllClick` of
`SkyDataViewConfig` are assigned more specific types.