feat(components/datetime): convert a user-entered digit into a date for the datepicker component (#179)
Datepicker numeric input is now translated to a date in the current month if the input is within the
current month's number of days.

BREAKING CHANGE: Datepicker numeric input is now translated to a date in the current month if the
input is within the current month's number of days. Numeric input outside of the current month's
number of days is now treated as invalid and is not converted to a `Date` object.