feat(components/tabs): remove support for not using a finish nav button
The nav buttons no longer check for the presence of a finish button when determining styling classes
and button visibility.

BREAKING CHANGE: This change removes support for not using a finish navigation button with the
previous and next wizard navigation buttons. To address this change, remove other save or finish
buttons and use the `sky-tabset-nav-button` of type `finish` instead.