feat(components/tabs): update the `SkyTabsetNavButtonComponent` `buttonType` input to no longer support `string` values (#1074)
BREAKING CHANGE: The tabset nav button component's `buttonType` input was set to allow values of type of `string` but it really only supported a handful of known `string` values represented by the `SkyTabsetNavButtonType` `string` union. This ability to specify a `string` value has been removed. This might cause problems if you are setting the `buttonType` input to a type of `string` in your consuming component's class.