feat(components/lists): update the `SkyRepeaterComponent` `expandMode` input to no longer support `string` values (#1076)
BREAKING CHANGE: The repeater component's `expandMode` input was set to allow values of type of `string` but it really only supported a handful of known `string` values represented by the `SkyRepeaterExpandModeType` `string` union. This ability to specify a `string` value has been removed. This might cause problems if you are setting the `expandMode` input to a type of `string` in your consuming component's class.