fix(components/packages): updated the `ng add` schematic to configure a workspace with a single project; workspaces with multiple projects should use the `add-skyux-to-project` schematic (#3659)
## Before this change

Consumers ran `ng add @skyux/packages --project my-app`, but requiring a
project name for the `add` schematic isn't an ideal pattern since an
`add` schematic is primarily used to install dependencies and add
configurations for the workspace. Most consumers would be accustomed to
running `ng add` without arguments, so the changes described below are
necessary.

## Summary of changes

Updated the `ng add` schematic to only configure a workspace if it
includes a single project. Workspaces with multiple projects should use
the new `add-skyux-to-project` schematic. This paradigm is used by
`angular-eslint` schematics as well.

Run for workspaces with a single project (likely an application):
```
ng add @skyux/packages
```

Run for each project in a multi-project workspace (likely a library and
a "showcase" application):
```
ng g @skyux/packages:add-skyux-to-project --project my-lib
ng g @skyux/packages:add-skyux-to-project --project my-lib-showcase
```