fix(sdk/eslint-config-skyux)!: re-enable `@angular-eslint/prefer-standalone` ESLint rule (#3837)
BREAKING CHANGE

Version 19 of `angular-eslint` introduced the
`@angular-eslint/prefer-standalone` rule, but it had a broken "fix"
implementation. Because of this, we disabled the rule entirely for SKY
UX 12 projects. The `angular-eslint` team has since removed the "fix"
implementation, so we can now re-enable this rule for our consumers in
SKY UX 13.