fix(components/autonumeric): return correct value when typing on an Android device (#4104)
[AB#3617801](https://dev.azure.com/blackbaud/f565481a-7bc9-4083-95d5-4f953da6d499/_workitems/edit/3617801)

This pull request is a workaround. ~I've created a [PR into
AutoNumeric](https://github.com/autoNumeric/autoNumeric/pull/821) to fix
the source, but we may need to merge this in if it doesn't work out.~
The "fix" felt like a hack. The AutoNumeric team recognizes that the
event handling is complex and [needs a
refactor](https://github.com/autoNumeric/autoNumeric/issues/400); let's
add this workaround until they've had a chance to clean things up on
their end.